Error message improvements (#4)
* Added support for secondary norminette command

Added a command2 and fileregex2 option to be able to run a secondary command, like norminette+.
Modified documentation to reflect the new feature.
Module now explicitly checks for each command if it is non-empty before it runs the command.
Code could be improved since there is a bunch of reused code, but it's functional...

* Update README.md

* removed h files from default for fileregex2 since norminette+ doesn't support .h files.

* massively improved extension.ts

     Updated: 2019/09/16 19:24:10 by lgutter       ########   odam.nl
Barely any copy paste code anymore. :)

* Fixed typo in README.

* Improvements in error display

Moved errors for empty lines to the top of the file since otherwise they are invisible.
Also 'fixed' the bold styling for errors.

* Fixed versioning

* Added gutter Icon

switched to a gutter icon indication errors, and empty line errors are back at their own line.
commented out the if statement for empty lines for now, since i'm hoping to figure out how to highlight the whole line just for the empty line errors.

* Fixed empty line error display

Moved empty line errors back to the correct line, but added a seperate DecorationType for empty lines that marks the whole line red.